Replace isViewDestroyed to getIsViewDestroyed in enqueueApiCall of ServerUtil
The problem of the old way where pass a boolean value is that the
function just refer the 'not-changeable' boolean value which is always
true if the function called when isViewDestroyed equals true.
So, I used the getter function for parameter so that We can get
the 'recent' isViewDestroyed successfully.